Make the Host Environment Variable Optional
Locust will allow authors to override the host by specifying one in the Locust class definition. However, the ability
to override this at runtime works only when no hostname has been specified on the command line. Even if the command
line's value is blank, the presence of the option still blocks the runtime customization.

This PR allows one to omit the environment variable for the host and thus specify hosts within their scripts. If no
environment variable is present, it simply skips adding the flag at startup.